+++ src/inc/InputHandler.h
@@ -5,12 +5,14 @@
 #include "ArgHandler.h"
 #include "TCPClient.h"
 #include "UDPClient.h"
-
 #include <iostream>
 #include <string>
 #include <sstream>
-#include <thread>
+#include <chrono>
+#include <sys/select.h>
+#include <unistd.h>
 #include <atomic>
+#include <thread>
 
 using namespace std;