+++ src/lib/TCPClient.cpp
@@ -1,13 +1,5 @@
 #include "../inc/TCPClient.h"
 
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <arpa/inet.h>
-#include <stdexcept>
-#include <iostream>
-#include <string>
-#include <cerrno>
-
 TCPClient::TCPClient(const ParsedArgs& args) :
     ProtocolClient(args.host, args.port) {
     printf_debug("TCPClient: Constructing...");